consolidate blur replacement integ config
cr: https://code.amazon.com/reviews/CR-63166952
