+++ .github/workflows/audio-video-transform-integration.yml
@@ -76,9 +76,6 @@
         - name: Run Background Blur Integration Test
           if:  steps.test_needed.outputs.integ_test_required == 'true'
           run: npm run test:integration-background-blur
-        - name: Run Background Replacement Integration Test
-          if:  steps.test_needed.outputs.integ_test_required == 'true'
-          run: npm run test:integration-background-replacement
         - name: Run Echo Reduction Integration Test
           if:  steps.test_needed.outputs.integ_test_required == 'true'
           run: npm run test:integration-echo-reduction
