--- integration/configs/background_replacement_test.config.json
@@ -1,36 +1,0 @@
-{
-  "`": 1,
-  "name": "background_replacement_test.config.json %ts",
-  "grids": [],
-  "tests": [
-    {
-      "name": "Background Replacement Test %ts",
-      "description": "Background Replacement Test",
-      "tupleSize": 1,
-      "noOfThreads": 1,
-      "testImpl": "BackgroundReplacementTest.js",
-      "payload": {
-        "url": "http://localhost:8080/",
-        "port": 30002,
-        "retry": 2
-      }
-    }
-  ],
-  "clients": [
-    {
-      "browserName": "chrome",
-      "version": "latest",
-      "platform": "MAC",
-       "video":{
-         "type":"Video",
-         "directory":"/Users/chef/",
-         "filename":"test_video"
-         },
-      "audio":{
-        "type":"Audio",
-        "directory":"/Users/chef/",
-        "filename":"test_video"
-        }
-    }
-  ]
-}
