+++ integration/configs/integrationTest/background_blur_test.config.json
@@ -14,6 +14,18 @@
         "port": 30002,
         "retry": 2
       }
+    },
+    {
+      "name": "Background Replacement Test %ts",
+      "description": "Background Replacement Test",
+      "tupleSize": 1,
+      "noOfThreads": 1,
+      "testImpl": "BackgroundReplacementTest.js",
+      "payload": {
+        "url": "http://localhost:8080/",
+        "port": 30002,
+        "retry": 2
+      }
     }
   ],
   "clients": [
